Remove PlayerVsPlayerInteractor and use DefaultGameInteractor
